Factored out the parallel-island code into guda.h.
This allows anyone to write their own genetic algorithm functions
and have the algorithm run on CUDA without having to dig into the
details of the parallel islands model. In addition, they need to
know almost nothing about how to write CUDA code.

Now asteroids.cu is an example of code that a user would write.